+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.42.0](https://github.com/bootstrap-vue-next/bootstrap-vue-next/compare/nuxt-v0.41.0...nuxt-v0.42.0) (2025-12-19)
+
+
+### Miscellaneous Chores
+
+* **nuxt:** Synchronize main group versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * bootstrap-vue-next bumped to 0.42.0
+  * peerDependencies
+    * bootstrap-vue-next bumped to 0.42.0
+
 ## [0.41.0](https://github.com/bootstrap-vue-next/bootstrap-vue-next/compare/nuxt-v0.40.9...nuxt-v0.41.0) (2025-12-12)